Project 2 RAG/requirements.txt
# Core dependencies
langchain>=0.3.0
langchain-community>=0.3.0
langchain-openai>=0.2.0
langchain-google-genai>=2.0.0
langchain-text-splitters>=0.3.0

# Vector store
faiss-cpu>=1.8.0

# Web search
tavily-python>=0.5.0

# Document loaders
pypdf>=5.0.0
wikipedia>=1.4.0

# UI
streamlit>=1.40.0

# Environment
python-dotenv>=1.0.0

# Utilities
pydantic>=2.5.0
tiktoken>=0.8.0
numpy>=1.26.0
